+++ src/JsonRpcX/Controllers/IJsonRpcController.cs
@@ -1,0 +1,9 @@
+namespace JsonRpcX.Controllers;
+
+/// <summary>
+/// Interface for tagging JSON RPC controllers.
+/// </summary>
+public interface IJsonRpcController
+{
+    // NOP
+}