+++ src/JsonRpcX/Methods/IJsonRpcMethodHandler.cs
@@ -1,9 +1,0 @@
-namespace JsonRpcX.Methods;
-
-/// <summary>
-/// Interface for tagging JSON RPC method handlers.
-/// </summary>
-public interface IJsonRpcMethodHandler
-{
-    // NOP
-}